More changes from size_t to uint64_t, use EXPECT_THROW_WITH_REGEX macro throughout config_impl_test.

+++ source/server/hot_restart_impl.h
@@ -48,7 +48,7 @@ private:
    * Initialize the shared memory segment, depending on whether we should be the first running
    * envoy, or a host restarted envoy process.
    */
-  static SharedMemory& initialize(size_t stats_set_size, Options& options);
+  static SharedMemory& initialize(uint64_t stats_set_size, Options& options);
 
   /**
    * Initialize a pthread mutex for process shared locking.
